requirements.txt
selenium>=4.0.0
python-dotenv
webdriver-manager
pdfplumber
prettytable
pytz
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pymupdf